Hotstar Clone
I developed a Hotstar clone, that replicates the popular features of
the original application. This project involved creating a user-friendly
interface and optimized the platform for both web and mobile use.
The project showcases my proficiency in front end development and
my ability to create a scalable and robust streaming platform